Altered IR class
Ryan, talking mainly to you.  I made a few changes to your class, feel
free to revert it if you don't like what I've done (yay version
control).  Basically, I've set it up so the class accepts the pin and
range info in the constructor so other calls don't need to know.  Also,
pretty much all the functions now return -1 as a return value for
errors.  Finally, I removed "Serial.begin(9600)" because that has
nothing to do with the IR sensors.
This was tested and it compiles in Arduino 1.0.3.