Make Claw class compile
Class wasn't compiling because it was missing the private members
leftServo and rightServo.  Also changed names to follow naming
conventions (leftservo - >leftServo, rightservo -> rightServo).  Added
private members leftPin and rightPin to represent which pins the servos
are attached to instead of using #defines (now the class doesn't have to
know).